int changed to size_t

+++ Project4/src/vm/swap.h
@@ -1,8 +1,9 @@
 #ifndef _SWAP_
 #define _SWAP_
-void swap_init (void);
-int swap_out (void *swap_pg);
-void swap_in (int index, void *swap_pg);
-void swap_free (int index);
+#include <stddef.h>
+void swap_init(void);
+size_t swap_out(void *swap_pg);
+void swap_in(size_t index, void *swap_pg);
+void swap_free(size_t index);
 
 #endif
